# src/backend/rest/api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: FridoLean.io
host: localhost:3001
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
...
paths:
...
  '/api/v1/projects/{projectID}/canvas/':
    x-swagger-router-controller: canvas
    post:
      tags:
        - project
        - canvas
      description: Create new Canvas
      operationId: createCanvas
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
          description: Bearer token
        - name: projectID
          in: path
          required: true
          type: string
          format: uuid
        - name: data
          required: true
          in: body
          description: Data of new Canvas
          schema:
            $ref: '#/definitions/UpdateCanvas'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/CanvasMeta'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/api/v1/projects/{projectID}/canvas/{canvasID}':
    x-swagger-router-controller: canvas
    get:
      tags:
        - project
        - canvas
      description: Gets details of Canvas
      operationId: readCanvas
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
          description: Bearer token
        - name: projectID
          in: path
          required: true
          type: string
          format: uuid
        - name: canvasID
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Canvas'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      tags:
        - project
        - canvas
      description: Delete Canvas
      operationId: deleteCanvas
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
          description: Bearer token
        - name: projectID
          in: path
          required: true
          type: string
          format: uuid
        - name: canvasID
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: Success
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
...
  '/api/v1/projects/{projectID}/canvas/{canvasID}/{buildingBlockType}':
    x-swagger-router-controller: buildingBlock
    post:
      tags:
        - project
        - entry
        - canvas
      description: Create new entry
      operationId: createEntry
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
          description: Bearer token
        - name: projectID
          in: path
          required: true
          type: string
          format: uuid
        - name: canvasID
          in: path
          required: true
          type: string
          format: uuid
        - name: buildingBlockType
          in: path
          description: Type of the building block you query (See models)
          required: true
          type: string
        - name: data
          in: body
          required: true
          description: Data of the new entry
          schema:
            properties:
              type:
                type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Entry'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/api/v1/projects/{projectID}/canvas/{canvasID}/{buildingBlockType}/{entryID}':
    x-swagger-router-controller: buildingBlock
    get:
      tags:
        - project
        - entry
        - canvas
      description: Gets content of entry
      operationId: readEntry
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
          description: Bearer token
        - name: projectID
          in: path
          required: true
          type: string
          format: uuid
        - name: canvasID
          in: path
          required: true
          type: string
          format: uuid
        - name: buildingBlockType
          in: path
          description: Type of the building block you query. (See models)
          required: true
          type: string
        - name: entryID
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Entry'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    put:
      tags:
        - project
        - entry
        - canvas
      description: Changes content of entry
      operationId: updateEntry
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
          description: Bearer token
        - name: projectID
          in: path
          required: true
          type: string
          format: uuid
        - name: canvasID
          in: path
          required: true
          type: string
          format: uuid
        - name: buildingBlockType
          in: path
          required: true
          type: string
        - name: entryID
          in: path
          required: true
          type: string
          format: uuid
        - name: data
          in: body
          required: true
          description: Data of the new entry
          schema:
            $ref: '#/definitions/UpdateEntry'
      responses:
        '200':
          description: Success
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      tags:
        - project
        - entry
        - canvas
      description: Deletes an entry. And returning its data.
      operationId: deleteEntry
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
          description: Bearer token
        - name: projectID
          in: path
          required: true
          type: string
          format: uuid
        - name: canvasID
          in: path
          required: true
          type: string
          format: uuid
        - name: buildingBlockType
          in: path
          required: true
          type: string
        - name: entryID
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: Success
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
...
definitions:
...
  UpdateEntry:
    properties:
      content:
        description: Content of the new Entry
        properties:
          title:
            type: string
          text:
            type: string
          target:
            type: string
          reference:
            type: string
  Entry:
    properties:
      _id:
        type: string
      type:
        type: string
      content:
        $ref: '#/definitions/Content'
  Content:
    properties:
      title:
        type: string
      text:
        type: string
      reference:
        type: string
      target:
        type: string
  BuildingBlock:
    properties:
      _id:
        type: string
        format: uuid
      title:
        type: string
        description: |
          BuisinessModelCanvas: [
            KeyPartners,
            KeyActivities,
            KeyResources,
            CustomerRelation,
            Channels,
            CustomerSegments,
            CostStructure,
            RevenueStreams,
            ValueProposition
          ],                                    

          ValuePropositionCanvas: [      
            ProductsAndServices,
            GainCreators,
            PainRelievers,
            Gains,
            Pains ,                                
            CustomerJobs
          ],                            

          LeanCanvas: [
            Problem,                                  
            Solution,                                   
            KeyMetrics,                                      
            UniqueValueProposition,                                                   
            UnfairAdvantages,                                           
            Channels,                                   
            CustomerSegments,                                           
            CostStructure,                                         
            RevenueStreams
          ],                                 

          CustomerJourney: [
            Advertisement,                                        
            PreSocialMedia,                                              
            PreWordOfMouth,                                               
            Expectations,                                       
            PastExperiences,                                           
            ServiceJourney,                                          
            Experiences,                                      
            RelationshipMenage,                                                  
            PostSocialMedia,                                               
            PostWordOfMouth,                                                
            DisSatisfaction
          ]
      buildingBlockType:
        type: string
      layoutEntry:
        properties:
          _id:
            type: string
            format: uuid
          x:
            type: number
          'y':
            type: number
          width:
            type: number
          height:
            type: number
      entries:
        type: array
        items:
          $ref: '#/definitions/Entry'
...
  UpdateCanvas:
    properties:
      title:
        type: string
      type:
        type: string
  Canvas:
    properties:
      _id:
        type: string
        format: uuid
      title:
        type: string
      canvasType:
        type: string
        description: BUSINESS_MODEL || LEAN || VALUE_PROPOSITION || CUSTOMER_JOURNEY
      lastEdited:
        type: integer
        format: int64
        description: 'Milliseconds since Unix epoch (Jan 1, 1970)'
      buildingBlocks:
        type: array
        items:
          $ref: '#/definitions/BuildingBlock'
  CanvasMeta:
    properties:
      _id:
        type: string
        format: uuid
      title:
        type: string
      type:
        type: string
      lastEdited:
        type: integer
        format: int64
        description: 'Milliseconds since Unix epoch (Jan 1, 1970)'
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...